refactor: move pipeline context tests into test_fix_pipeline
Tests belong alongside other pipeline tests since they exercise the
same entry point (run_fix_pipeline). Removes the separate
test_pipeline_service.py file and reuses existing mock helpers.
